M3: Extract AST introspection helpers to utils/ast_inspection.py
Move `_get_func_indexing_params`, `_slice_references_params`,
`_collect_subscripts`, `_extract_bare_names` from
`utils/error_handling.py` into a new focused module. Update imports
in `pandas_utils.py` and `tests/test_validate_array_indexing.py`.
`error_handling.py::validate_transition_probs` still uses
`_get_func_indexing_params` and now imports it from the new module
(deferred to M5' for full extraction). Part of Phase 1 — see
`Phase 1 — Validation Cleanup.md`.
